+++ tests/test_httpbroker.py
@@ -209,6 +209,33 @@ class GetFunctionTests(mocker.MockerTestCase):
             lambda: httpbroker.get('http://manager.scielo.org/api/v1/', resource_id='70')
         )
 
+    def test_https_turns_off_ca_cert_verification(self):
+        import requests
+        mock_response = self.mocker.mock(requests.Response)
+        mock_response.json()
+        self.mocker.result({'title': 'foo'})
+        mock_response.status_code
+        self.mocker.result(200)
+
+        mock_requests_get = self.mocker.mock()
+        mock_requests_get('https://manager.scielo.org/api/v1/journals/70/',
+                          headers=mocker.ANY,
+                          params=None,
+                          verify=False)
+        self.mocker.result(mock_response)
+
+        mock_requests = self.mocker.replace('requests')
+        mock_requests.get
+        self.mocker.result(mock_requests_get)
+
+        self.mocker.replay()
+
+        self.assertEqual(
+            httpbroker.get('https://manager.scielo.org/api/v1/',
+                endpoint='journals', resource_id='70'),
+            {'title': 'foo'}
+        )
+
 
 class PostFunctionTests(mocker.MockerTestCase):
 
@@ -307,4 +334,9 @@ class MakeFullUrlFunctionTests(unittest.TestCase):
         path_segments = ['manager.scielo.org', 'api', 'v1', 'journals']
         self.assertEqual(httpbroker._make_full_url(*path_segments),
             'http://manager.scielo.org/api/v1/journals/')
+
+    def test_https(self):
+        path_segments = ['https://manager.scielo.org', 'api', 'v1', 'journals']
+        self.assertEqual(httpbroker._make_full_url(*path_segments),
+            'https://manager.scielo.org/api/v1/journals/')