+++ HISTORY.md
@@ -5,6 +5,7 @@ History
 ---
 
 * Added `tox.ini` to help the porting to Python3.3.
+* Support for https (without verifiying CA)
 
 
 0.4 (2013-08-30)