+++ HISTORY.md
@@ -1,12 +1,13 @@
 History
 =======
 
-0.5
----
+0.5 (2014-02-10)
+----------------
 
 * Added `tox.ini` to help the porting to Python3.3.
 * Support for https (without verifiying CA).
 * Added `Content-Type: application/json` HTTP header to all post requests.
+* Added a do-nothing logger handler by default.
 
 
 0.4 (2013-08-30)